pyproject.toml
[tool.poetry]
name = "eleva-project"
version = "0.1.0"
description = "Technical Assignment Elevatus Company"
authors = ["Yousef Aldabbas <yousefaldabbas0@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.108.0"
alembic = "^1.13.1"
odmantic = "^1.0.0"
pydantic = {extras = ["email"], version = "^2.5.3"}
uvicorn = "^0.25.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
beanie = "^1.24.0"
bcrypt = "4.0.1"
cloudinary = "^1.37.0"
isort = "^5.13.2"
asgi-correlation-id = "^4.3.0"
ddtrace = "^2.4.0"
structlog = "^23.3.0"
pandas = "^2.1.4"
pydantic-settings = "^2.1.0"
asgi-lifespan = "^2.1.0"
httpx = "^0.26.0"
pytest = "^7.4.4"
anyio = "^4.2.0"
pytest-asyncio = "^0.23.3"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^23.12.1"
pytest-ordering = "^0.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
